// frontend/src/api-client/models/PlaylistClip.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Reelshelf | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
import type { Clip } from "./Clip";
import {
  ClipFromJSON,
  ClipFromJSONTyped,
  ClipToJSON,
  ClipToJSONTyped,
} from "./Clip";

/**
 *
 * @export
 * @interface PlaylistClip
 */
export interface PlaylistClip {
  /**
   *
   * @type {string}
   * @memberof PlaylistClip
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof PlaylistClip
   */
  clipId: string;
  /**
   *
   * @type {number}
   * @memberof PlaylistClip
   */
  position: number;
  /**
   *
   * @type {string}
   * @memberof PlaylistClip
   */
  addedByUserId: string;
  /**
   *
   * @type {Date}
   * @memberof PlaylistClip
   */
  addedAt: Date;
  /**
   *
   * @type {Clip}
   * @memberof PlaylistClip
   */
  clipDetails?: Clip | null;
}

/**
 * Check if a given object implements the PlaylistClip interface.
 */
export function instanceOfPlaylistClip(value: object): value is PlaylistClip {
  if (!("id" in value) || value["id"] === undefined) return false;
  if (!("clipId" in value) || value["clipId"] === undefined) return false;
  if (!("position" in value) || value["position"] === undefined) return false;
  if (!("addedByUserId" in value) || value["addedByUserId"] === undefined)
    return false;
  if (!("addedAt" in value) || value["addedAt"] === undefined) return false;
  return true;
}

export function PlaylistClipFromJSON(json: any): PlaylistClip {
  return PlaylistClipFromJSONTyped(json, false);
}

export function PlaylistClipFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): PlaylistClip {
  if (json == null) {
    return json;
  }
  return {
    id: json["id"],
    clipId: json["clip_id"],
    position: json["position"],
    addedByUserId: json["added_by_user_id"],
    addedAt: new Date(json["added_at"]),
    clipDetails:
      json["clip_details"] == null
        ? undefined
        : ClipFromJSON(json["clip_details"]),
  };
}

export function PlaylistClipToJSON(json: any): PlaylistClip {
  return PlaylistClipToJSONTyped(json, false);
}

export function PlaylistClipToJSONTyped(
  value?: PlaylistClip | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    id: value["id"],
    clip_id: value["clipId"],
    position: value["position"],
    added_by_user_id: value["addedByUserId"],
    added_at: value["addedAt"].toISOString(),
    clip_details: ClipToJSON(value["clipDetails"]),
  };
}